[WIP] cleanup: remove runtime CPU detection to enable automatic building
postgres is one of the formulae which supports runtime CPU detection, per

https://github.com/Homebrew/homebrew-core/blob/master/style_exceptions/runtime_cpu_detection_allowlist.json

Per https://github.com/orgs/Homebrew/discussions/4684#discussioncomment-6599218,
if CPU runtime detection is enabled, there will be a check that
the package name is included in the aforementioned allowlist.json.

Because of this, our tap/fork of the postgresql package is not automatically building.

Let's see if removing this field will allow automatic builds.